requirements.txt
tensorflow-gpu~=2.9.1
keras~=2.9.0
pandas~=1.4.3
matplotlib~=3.5.3
numpy~=1.23.2